refactor: Use Appium as peer dependency
