+++ lib/commands/file-movement.js
@@ -1,7 +1,7 @@
 import _ from 'lodash';
 import path from 'path';
-import { errors } from '@appium/base-driver';
-import { fs, mkdirp, util, zip } from '@appium/support';
+import { errors } from 'appium/driver';
+import { fs, mkdirp, util, zip } from 'appium/support';
 import { MODIFY_FS_FEATURE } from '../constants';
 
 // List of env variables, that can be expanded in path
