+++ lib/commands/find.js
@@ -1,4 +1,4 @@
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 
 
 const commands = {};
