+++ lib/commands/general.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 
 async function getScreenSize (log) {
   const dimensions = await this.execPowerShell({
