+++ lib/commands/powershell.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { fs, tempDir } from '@appium/support';
+import { fs, tempDir } from 'appium/support';
 import { exec } from 'teen_process';
 import path from 'path';
 import B from 'bluebird';
