+++ lib/commands/record-screen.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
 import { waitForCondition } from 'asyncbox';
-import { util, fs, net, system, tempDir } from '@appium/support';
+import { util, fs, net, system, tempDir } from 'appium/support';
 import { SubProcess } from 'teen_process';
 import B from 'bluebird';
