+++ lib/driver.js
@@ -1,6 +1,6 @@
 import _ from 'lodash';
-import { BaseDriver } from '@appium/base-driver';
-import { system } from '@appium/support';
+import { BaseDriver } from 'appium/driver';
+import { system } from 'appium/support';
 import { WinAppDriver } from './winappdriver';
 import { desiredCapConstraints } from './desired-caps';
 import commands from './commands/index';
