+++ lib/installer.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import { system, fs, util, net, tempDir } from '@appium/support';
+import { system, fs, util, net, tempDir } from 'appium/support';
 import path from 'path';
 import { exec } from 'teen_process';
 import log from './logger';
