+++ lib/winappdriver.js
@@ -1,12 +1,12 @@
 import _ from 'lodash';
 import os from 'os';
 import path from 'path';
-import { JWProxy, errors } from '@appium/base-driver';
+import { JWProxy, errors } from 'appium/driver';
 import { SubProcess } from 'teen_process';
 import { getWADExecutablePath } from './installer';
 import { waitForCondition } from 'asyncbox';
 import { execSync } from 'child_process';
-import { util } from '@appium/support';
+import { util } from 'appium/support';
 import { findAPortNotInUse } from 'portscanner';
