+++ package.json
@@ -41,8 +41,6 @@
     "build/lib"
   ],
   "dependencies": {
-    "@appium/base-driver": "^8.3.0",
-    "@appium/support": "^2.55.3",
     "@babel/runtime": "^7.0.0",
     "asyncbox": "^2.3.1",
     "bluebird": "^3.5.1",
@@ -71,11 +69,16 @@
     "precommit-msg",
     "precommit-test"
   ],
+  "peerDependencies": {
+    "appium": "^2.0.0-beta.40"
+  },
   "devDependencies": {
     "@appium/gulp-plugins": "^7.0.0",
     "@appium/eslint-config-appium": "^6.0.0",
     "@appium/test-support": "^1.0.0",
     "@semantic-release/git": "^10.0.1",
+    "appium": "^2.0.0-beta.40",
+    "appium-windows-driver": "file:.",
     "chai": "^4.1.2",
     "chai-as-promised": "^7.1.1",
     "eslint-config-prettier": "^8.5.0",
