+++ test/e2e/driver-e2e-specs.js
@@ -1,7 +1,7 @@
 import { remote as wdio } from 'webdriverio';
 import chai from 'chai';
 import chaiAsPromised from 'chai-as-promised';
-import { startServer } from '../../lib/server';
+import { startServer } from '../server';
 import { isAdmin } from '../../lib/installer';
 
 chai.should();
@@ -12,7 +12,8 @@ const TEST_HOST = 'localhost';
 
 const TEST_CAPS = {
   platformName: 'Windows',
-  'appium:app': 'Microsoft.WindowsCalculator_8wekyb3d8bbwe!App'
+  'appium:automationName': 'windows',
+  'appium:app': 'Microsoft.WindowsCalculator_8wekyb3d8bbwe!App',
 };
 
 const WDIO_OPTS = {
